img-1130 - try again to fix scipy incompatibility
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@10463 1fc53939-2000-0410-845c-e8453a809027

+++ cellprofiler/cpmath/cpmorphology.py
@@ -354,9 +354,9 @@ def convex_hull_ijv(pixel_labels, indexes):
     # Calculate the centers for each label
     #
     center_i = fixup_scipy_ndimage_result(
-        scind.mean(i.astype(float), labels_per_point, indexes))
+        scind.mean(i.astype(float), labels_per_point.astype(int), indexes))
     center_j = fixup_scipy_ndimage_result(
-        scind.mean(j.astype(float), labels_per_point, indexes))
+        scind.mean(j.astype(float), labels_per_point.astype(int), indexes))
     centers = np.column_stack((center_i, center_j))
     #
     # Now make an array with one outline point per row and the following
